Reduce number of memmove and fread calls
Instead of using memmove to set the 'memmem' match ptr to the start of
the buffer I now use a pointer to keep track what we've read in the
buffer.  The only time I call memmove and fread is when the buffer get's
too small, or when we can't find the seiString within the buffer.